Extract WaitForAPI component
