+++ packages/ui/src/app/index.ts
@@ -1,1 +1,2 @@
 export * from './App'
+export { WaitForAPI } from '@/app/components/WaitForAPI'
